Decided to create new entities
Art pending and need to learn how to implement, but decided to create another Zombie and another plant to be used. For now the commented part i added to main was to implement but VSC will not properly import atm so i will just push this for now. Will work with more later